Refactor: get correct member data for reputation dialog
