+++ src/utils/members.ts
@@ -1,0 +1,10 @@
+import { Member } from '~types';
+import { notNull, notUndefined } from '~utils/arrays';
+
+export const extractUsersFromColonyMemberData = (
+  members: Member[] | null | undefined,
+) =>
+  members
+    ?.map((member) => member.user)
+    .filter(notNull)
+    .filter(notUndefined) || [];
